requirements.txt
Flask==1.0.2
pyyaml==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
google-cloud-kms==1.2.1
google-cloud-storage==1.23.0
